Created BrandInput component and refactored AddToWishlistModal so all logic involving the brand input is now in the new component so it can be used elsewhere.
