+++ src/components/forms/EditGear.jsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import { useNavigate, useParams } from "react-router-dom";
 import { editGearPost, getPostById } from "../services/postService";
 import { GearForm } from "./GearForm";
-import { DropdownUtility } from "../dropdowns/DropdownUtility.jsx";
+import { DropdownUtility } from "../utilities/DropdownUtility.jsx";
 
 export const EditGear = ({ currentUser }) => {
   const [initialData, setInitialData] = useState(null);
